It is possible to convert bool into Numeric and vice versa in sqlite/go

+++ create_struct.sql
@@ -13,7 +13,7 @@ CREATE TABLE products(
     name  TEXT NOT NULL,
     product_id TEXT NOT NULL UNIQUE,
     category TEXT NOT NULL,
-    quanto INTEGER NOT NULL,
+    quanto NUMERIC NOT NULL,
     creationDate TEXT NOT NULL,
     expirationDate TEXT NOT NULL
 );
